fix: constants not limited to max int range

+++ Team14/Code14/src/spa/src/commons/parser/expr_definition/expr_grammar.cpp
@@ -88,7 +88,7 @@ VariableNode *VariableGrammarRule::parseNode(TokenIterator &tokenStream) {
 
 ConstantNode *ConstantGrammarRule::parseNode(TokenIterator &tokenStream) {
   if ((*tokenStream)->type == TokenType::kLiteral) {
-    return new ConstantNode(std::stoi((*tokenStream++)->value));
+    return new ConstantNode(std::stoll((*tokenStream++)->value));
   }
   throw ParseSyntaxError("Expected constant, got " + (*tokenStream++)->value);
 }
